+++ Kaeon FUSION/Interfaces/Standard/Source/standard_kaeon_fusion/commands/math/Multiply.java
@@ -25,6 +25,6 @@ public class Multiply extends FUSIONUnit {
 		for(int i = 1; i < processed.size(); i++)
 			result *= Double.parseDouble("" + processed.get(i));
 		
-		return result % 1 != 0 ? result : "" + (int) result;
+		return result % 1 != 0 ? result : "" + (long) result;
 	}
 }
